refactor: Разбей скрипты на модули
